Day 2 questions
I did question 1 approach but question 88 was a bit tricky I approached it differently and didn't read the question correctly so wasted some time there. After reading the question correctly realised its a bit hard for me to solve so looked the discussions and found 1 solution. I went through it and it was nicely explained but I'm too tired to write this approach as it looks lengthy, hopefully you can write it or any other approach you feel can be used here.